Fix type mismatch in onGetReport when-expression
REPORT_TYPE_INPUT/FEATURE are Java byte constants (Kotlin Byte).
Comparing type.toInt() (Int) against them caused compile errors.
Drop .toInt() so both sides of the when are Byte.

+++ app/src/main/kotlin/com/rishabh/btgamepad/hid/BluetoothHidService.kt
@@ -233,7 +233,7 @@ class BluetoothHidService : Service() {
 
         override fun onGetReport(device: BluetoothDevice, type: Byte, id: Byte, bufferSize: Int) {
             val hid = hidDevice ?: return
-            when (type.toInt()) {
+            when (type) {
                 BluetoothHidDevice.REPORT_TYPE_INPUT -> {
                     if (id.toInt() == HidConstants.REPORT_ID)
                         hid.replyReport(device, type, id, reportBuilder.toReport())
